Reorganize and define base SQLAlchemy backend API
