+++ ibis/backends/base_sql/compiler.py
@@ -1,7 +1,6 @@
 import ibis.backends.base_sqlalchemy.compiler as comp
 import ibis.expr.operations as ops
 from ibis.backends.base.sql import (
-    binary_infix_ops,
     operation_registry,
     quote_identifier,
 )
@@ -87,14 +86,11 @@ class BaseTableSetFormatter(comp.TableSetFormatter):
         return quote_identifier(name)
 
 
-_operation_registry = {**operation_registry, **binary_infix_ops}
-
-
 # TODO move the name method to comp.ExprTranslator and use that instead
 class BaseExprTranslator(comp.ExprTranslator):
     """Base expression translator."""
 
-    _registry = _operation_registry
+    _registry = operation_registry
     context_class = BaseContext
 
     @staticmethod
