+++ ibis/backends/mysql/__init__.py
@@ -1,5 +1,5 @@
 from ibis.backends.base import BaseBackend
-from ibis.backends.base_sqlalchemy.alchemy import AlchemyQueryBuilder
+from ibis.backends.base.sql.alchemy import AlchemyQueryBuilder
 
 from .client import MySQLClient, MySQLDatabase, MySQLTable
 from .compiler import MySQLExprTranslator
