+++ ibis/backends/mysql/client.py
@@ -6,7 +6,8 @@ import pymysql  # NOQA fail early if the driver is missing
 import sqlalchemy as sa
 import sqlalchemy.dialects.mysql as mysql
 
-import ibis.backends.base_sqlalchemy.alchemy as alch
+from ibis.backends.base.sql.alchemy import (
+    AlchemyTable, AlchemyDatabaseSchema, AlchemyDatabase, AlchemyClient)
 import ibis.expr.datatypes as dt
 
 # TODO(kszucs): unsigned integers
@@ -32,19 +33,19 @@ def mysql_blob(satype, nullable=True):
     return dt.Binary(nullable=nullable)
 
 
-class MySQLTable(alch.AlchemyTable):
+class MySQLTable(AlchemyTable):
     pass
 
 
-class MySQLSchema(alch.AlchemyDatabaseSchema):
+class MySQLSchema(AlchemyDatabaseSchema):
     pass
 
 
-class MySQLDatabase(alch.AlchemyDatabase):
+class MySQLDatabase(AlchemyDatabase):
     schema_class = MySQLSchema
 
 
-class MySQLClient(alch.AlchemyClient):
+class MySQLClient(AlchemyClient):
 
     """The Ibis MySQL client class
