+++ ibis/backends/postgres/__init__.py
@@ -1,6 +1,6 @@
 """PostgreSQL backend."""
 from ibis.backends.base import BaseBackend
-from ibis.backends.base_sqlalchemy.alchemy import AlchemyQueryBuilder
+from ibis.backends.base.sql.alchemy import AlchemyQueryBuilder
 
 from .client import PostgreSQLClient, PostgreSQLDatabase, PostgreSQLTable
 from .compiler import PostgreSQLExprTranslator
