+++ ibis/backends/sqlite/__init__.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 from ibis.backends.base import BaseBackend
-from ibis.backends.base_sqlalchemy.alchemy import AlchemyQueryBuilder
+from ibis.backends.base.sql.alchemy import AlchemyQueryBuilder
 
 from .client import SQLiteClient, SQLiteDatabase, SQLiteTable
 from .compiler import SQLiteExprTranslator
