+++ ibis/backends/sqlite/client.py
@@ -8,11 +8,11 @@ from typing import Optional
 import regex as re
 import sqlalchemy as sa
 
-import ibis.backends.base_sqlalchemy.alchemy as alch
+from ibis.backends.base.sql.alchemy import AlchemyTable, AlchemyClient
 from ibis.client import Database
 
 
-class SQLiteTable(alch.AlchemyTable):
+class SQLiteTable(AlchemyTable):
     pass
 
 
@@ -311,7 +311,7 @@ def _register_aggregate(agg, con):
     con.connection.connection.create_aggregate(agg.__name__, nargs, agg)
 
 
-class SQLiteClient(alch.AlchemyClient):
+class SQLiteClient(AlchemyClient):
     """The Ibis SQLite client class."""
 
     def __init__(self, backend, path=None, create=False):
